requirements.txt
discord.py >= 0.16
youtube-dl >= 2017.4.17
requests >= 2.18.4